feat(cli): add html export format, unify serialization dispatch, and rename format flag
